Code for Module 2 Multiple stock year data.bas
This is my code for the file (multiple stock year data) as I understand the code to work. I reached out to instructor and Ask BCS as code took a long time to run but works with a smaller form of the data. Please see AskBCS file I saved in the repo as well for the code they edited and provided to optimize the code.